+++ CHANGELOG.md
@@ -23,6 +23,7 @@ You can also check the
   - Hover on operator list for sunshine indicators highlights operators on map
   - Sunshine map has better styling and is closer to the current prices map
   - Added Latest Year indicator to map details panel mini line chart
+  - Added support for visualizing Dynamic Tariffs, data support to come in the future
 
 - Fixes